pass grade feeze info to xblocks
